+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/util/PaginationFilterSink.java
@@ -71,4 +71,6 @@ public interface PaginationFilterSink<T> {
 
   public void restrictPaginationByUser(T item, String userName, boolean creator, Consumer<String> errorHandler);
 
+  public void restrictPaginationByUserOrGroup(T item, String name, boolean creator, Consumer<String> errorHandler);
+
 }