+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@ Changes:
     before/after specified range
   * Search for items by ID
   * Search worksets by user, date, and ID
+  * Search by user group in any way that you could previously search by user
   * Fix bug where unassigned JIRA tickets will cause tables to not display correctly
 
 Upgrade notes: